requirements.txt
mido
python-rtmidi
dbus-python
PyGObject
